implemented id Gamma once talking to cyrus and figuring out what it actually is. silly mistake.
adding a proof that equality of htype is decidable. we need this (at least)
in expandability, to pick between ESAsc1 and ESAsc2. it means that if we
expand to a language where types aren't deciable then we'll have some
problem -- @cyrus- says that if you can't prove that two types are equal,
in the gradual typing work, you'd just err on the side of caution, assume
they aren't, and insert the cast to be safe. that's not quite the notion
that's going on here, but worth remembering.

did most of a case of expandability; some holes there that i'm not sure how
to fix