typo in the expandablity ana statement with => instead of <= in the typing premise. fixed that, which made a problem go away.
@cyrus- : i changed the format of the ESVar rule to be

```
 ∀{Γ x τ} → (x , τ) ∈ Γ →
                         Γ ⊢ X x ⇒ τ ~> X x ⊣ ∅
```
instead of

```
 ∀{Γ x τ} → (Γ ,, (x , τ)) ⊢ X x ⇒ τ ~> X x ⊣ ∅
```
this doesn't match what you wrote in the math text as on the nose, but is
consistent with SVar, which is formulated as in the POPL work from last
year. let me know what you think about this. it mostly just saves a really
pointless argument about _where_ in the context the binding appears.