i commented out the disjointness premises i had in the rules, pending thinking about how to actually do this.
one way to figure it out is to just see where it fails in later proofs,
probably preservation?, and then come back and try to figure out what the
least we need to say is to let those proofs go through. there may be a
better datastructure or some augmented form of what i use now that means
that the contexts themselves carry around the information needed, rather
than adding a premise.

i definitely want to avoid cluttering up the statements of the theorems
with context crap all over the place, which is a real risk