these changes are the result of working with @cyrus- yesterday on preservation for an hour or so. his idea was that it might be easier to prove preservation for the instruction transition judgement and then prove that evaluation contexts respect it; that's reflected in the handful of new lemmas in that file.
we also now suspect that there is an error in the ITCast rule; there are a
few places throughout where we need two types to be the same but because of
that rule only know that they are related by ~; one of these is in one of
the preservation lemmas but i've seen it elsewhere. @cyrus- is looking into
some of the other work on gradaual typing to see if we're following the
right path.