work towards #5
talking to cyrus yesterday, we added a premise for finality in the ap error
case: we're not interested in errors that aren't about to happen, or that
don't correspond to steps that might be taken. this kills off one of the two counter
examples with a bit of monster barring.

this means that the corresponding case goes through, with a side condition
about evaluation contexts and finality.